refactor(amd64): deduplicate codegen lowering
